requirements.txt
eel
bottle
psycopg2-binary
prometheus_client>=0.14.1
